# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "BarcodeOut": {
        "properties": {
          "draft": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ItemDraftOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "item": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ItemOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "enum": ["item", "draft", "not_found"],
            "title": "Status",
            "type": "string"
          }
        },
        "required": ["status", "item", "draft"],
        "title": "BarcodeOut",
        "type": "object"
      },
...
      "CopyDayIn": {
        "properties": {
          "source_day": {
            "format": "date",
            "title": "Source Day",
            "type": "string"
          }
        },
        "required": ["source_day"],
        "title": "CopyDayIn",
        "type": "object"
      },
      "CopyEntryIn": {
        "properties": {
          "day": {
            "format": "date",
            "title": "Day",
            "type": "string"
          },
          "slot": {
            "anyOf": [
              {
                "enum": ["breakfast", "lunch", "dinner", "snack"],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Slot"
          }
        },
        "required": ["day"],
        "title": "CopyEntryIn",
        "type": "object"
      },
      "DayOut": {
        "properties": {
          "day": {
            "format": "date",
            "title": "Day",
            "type": "string"
          },
          "people": {
            "items": {
              "$ref": "#/components/schemas/PersonDayOut"
            },
            "title": "People",
            "type": "array"
          }
        },
        "required": ["day", "people"],
        "title": "DayOut",
        "type": "object"
      },
      "DayTypeIn": {
        "properties": {
          "day_type": {
            "anyOf": [
              {
                "enum": ["training", "rest"],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Day Type"
          }
        },
        "required": ["day_type"],
        "title": "DayTypeIn",
        "type": "object"
      },
      "EntryIn": {
        "properties": {
          "at": {
            "anyOf": [
              {
                "format": "time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "At"
          },
          "components": {
            "items": {
              "$ref": "#/components/schemas/ComponentIn"
            },
            "maxItems": 50,
            "title": "Components",
            "type": "array"
          },
          "day": {
            "format": "date",
            "title": "Day",
            "type": "string"
          },
          "eaten_out": {
            "default": false,
            "title": "Eaten Out",
            "type": "boolean"
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "portions": {
            "default": 1.0,
            "title": "Portions",
            "type": "number"
          },
          "saved_meal_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Saved Meal Id"
          },
          "slot": {
            "enum": ["breakfast", "lunch", "dinner", "snack"],
            "title": "Slot",
            "type": "string"
          }
        },
        "required": ["day", "slot"],
        "title": "EntryIn",
        "type": "object"
      },
      "EntryOut": {
        "properties": {
          "at": {
            "format": "time",
            "title": "At",
            "type": "string"
          },
          "components": {
            "items": {
              "$ref": "#/components/schemas/ComponentOut"
            },
            "title": "Components",
            "type": "array"
          },
          "day": {
            "format": "date",
            "title": "Day",
            "type": "string"
          },
          "eaten_out": {
            "title": "Eaten Out",
            "type": "boolean"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "intake": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TotalsOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "saved_meal_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Saved Meal Id"
          },
          "share": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Share"
          },
          "slot": {
            "enum": ["breakfast", "lunch", "dinner", "snack"],
            "title": "Slot",
            "type": "string"
          },
          "state": {
            "anyOf": [
              {
                "enum": ["planned", "logged", "skipped"],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "State"
          }
        },
        "required": [
          "id",
          "day",
          "slot",
          "at",
          "name",
          "eaten_out",
          "saved_meal_id",
          "state",
          "share",
          "components",
          "intake"
        ],
        "title": "EntryOut",
        "type": "object"
      },
      "EntryPatchIn": {
        "properties": {
          "at": {
            "anyOf": [
              {
                "format": "time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "At"
          },
          "components": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ComponentIn"
                },
                "maxItems": 50,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Components"
          },
          "day": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Day"
          },
          "eaten_out": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Eaten Out"
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "slot": {
            "anyOf": [
              {
                "enum": ["breakfast", "lunch", "dinner", "snack"],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Slot"
          }
        },
        "title": "EntryPatchIn",
        "type": "object"
      },
      "EntryStateIn": {
        "properties": {
          "state": {
            "enum": ["planned", "logged", "skipped"],
            "title": "State",
            "type": "string"
          }
        },
        "required": ["state"],
        "title": "EntryStateIn",
        "type": "object"
      },
      "FavouriteIn": {
        "properties": {
          "on": {
            "title": "On",
            "type": "boolean"
          }
        },
        "required": ["on"],
        "title": "FavouriteIn",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ItemOut": {
        "properties": {
          "barcodes": {
            "items": {
              "type": "string"
            },
            "title": "Barcodes",
            "type": "array"
          },
          "base_unit": {
            "enum": ["g", "ml"],
            "title": "Base Unit",
            "type": "string"
          },
          "brand": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Brand"
          },
          "category": {
            "title": "Category",
            "type": "string"
          },
          "favourite": {
            "title": "Favourite",
            "type": "boolean"
          },
          "generic": {
            "title": "Generic",
            "type": "boolean"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Url"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "names": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ]
            },
            "title": "Names",
            "type": "object"
          },
          "nutrients": {
            "$ref": "#/components/schemas/NutrientsOut"
          },
          "package_size": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Package Size"
          },
          "servings": {
            "items": {
              "$ref": "#/components/schemas/ServingOut"
            },
            "title": "Servings",
            "type": "array"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "tracking_mode": {
            "enum": ["counted", "status"],
            "title": "Tracking Mode",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "names",
          "brand",
          "category",
          "base_unit",
          "nutrients",
          "package_size",
          "servings",
          "barcodes",
          "tracking_mode",
          "source",
          "generic",
          "favourite",
          "image_url"
        ],
        "title": "ItemOut",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Mahlzeit",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/barcodes/{code}": {
      "get": {
        "operationId": "lookup_barcode_api_barcodes__code__get",
        "parameters": [
          {
            "in": "path",
            "name": "code",
            "required": true,
            "schema": {
              "title": "Code",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BarcodeOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Lookup Barcode",
        "tags": ["catalogue"]
      }
    },
...
    "/api/days/{day_}": {
      "get": {
        "operationId": "get_day_api_days__day___get",
        "parameters": [
          {
            "in": "path",
            "name": "day_",
            "required": true,
            "schema": {
              "format": "date",
              "title": "Day ",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DayOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Day",
        "tags": ["day"]
      }
    },
    "/api/days/{day_}/copy": {
      "post": {
        "operationId": "copy_day_api_days__day___copy_post",
        "parameters": [
          {
            "in": "path",
            "name": "day_",
            "required": true,
            "schema": {
              "format": "date",
              "title": "Day ",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CopyDayIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/EntryOut"
                  },
                  "title": "Response Copy Day Api Days  Day   Copy Post",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Copy Day",
        "tags": ["day"]
      }
    },
    "/api/days/{day_}/day-type": {
      "put": {
        "operationId": "set_day_type_api_days__day___day_type_put",
        "parameters": [
          {
            "in": "path",
            "name": "day_",
            "required": true,
            "schema": {
              "format": "date",
              "title": "Day ",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DayTypeIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Day Type",
        "tags": ["day"]
      }
    },
    "/api/entries": {
      "post": {
        "operationId": "log_food_api_entries_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EntryIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntryOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Log Food",
        "tags": ["day"]
      }
    },
    "/api/entries/{entry_id}": {
      "delete": {
        "operationId": "delete_entry_api_entries__entry_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "entry_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Entry Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Entry",
        "tags": ["day"]
      },
      "patch": {
        "operationId": "update_entry_api_entries__entry_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "entry_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Entry Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EntryPatchIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntryOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Entry",
        "tags": ["day"]
      }
    },
    "/api/entries/{entry_id}/copy": {
      "post": {
        "operationId": "copy_entry_api_entries__entry_id__copy_post",
        "parameters": [
          {
            "in": "path",
            "name": "entry_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Entry Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CopyEntryIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntryOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Copy Entry",
        "tags": ["day"]
      }
    },
...
    "/api/entries/{entry_id}/state": {
      "post": {
        "operationId": "set_entry_state_api_entries__entry_id__state_post",
        "parameters": [
          {
            "in": "path",
            "name": "entry_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Entry Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EntryStateIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntryOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Entry State",
        "tags": ["day"]
      }
    },
...
    "/api/items": {
      "get": {
        "operationId": "search_items_api_items_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "default": "",
              "maxLength": 100,
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 50,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ItemOut"
                  },
                  "title": "Response Search Items Api Items Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search Items",
        "tags": ["catalogue"]
      },
      "post": {
        "operationId": "create_item_api_items_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ItemIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Item",
        "tags": ["catalogue"]
      }
    },
    "/api/items/{item_id}": {
      "get": {
        "operationId": "get_item_api_items__item_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "item_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Item Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Item",
        "tags": ["catalogue"]
      },
      "put": {
        "operationId": "update_item_api_items__item_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "item_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Item Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ItemIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Item",
        "tags": ["catalogue"]
      }
    },
    "/api/items/{item_id}/favourite": {
      "put": {
        "operationId": "set_favourite_api_items__item_id__favourite_put",
        "parameters": [
          {
            "in": "path",
            "name": "item_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Item Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FavouriteIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Favourite",
        "tags": ["catalogue"]
      }
    },
...
    "/api/online-search": {
      "get": {
        "operationId": "search_online_api_online_search_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "maxLength": 100,
              "minLength": 2,
              "title": "Q",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ItemDraftOut"
                  },
                  "title": "Response Search Online Api Online Search Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search Online",
        "tags": ["catalogue"]
      }
    },
...
  }
}